# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stories API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/story-items/{story_item_id}/story-item-buttons": {
            "get": {
                "tags": [
                    "Story"
                ],
                "summary": "Получение списка кнопок для элемента сториса",
                "operationId": "e5d5cf559a1fe20d3abe86d3f90b3950",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "story_item_id",
                        "in": "path",
                        "description": "ID элемента сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 21
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "position"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "button_text": {
                                                        "type": "string",
                                                        "example": "Yes"
                                                    },
                                                    "media_url": {
                                                        "type": "string",
                                                        "example": "http://stories.test/storage/stories/items/buttons/01J7DK35EP99WJ4VBXMV3HN48Q.webp"
                                                    },
                                                    "position": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apiKey",
                "in": "header"
            }
        }
    },
...
}